Add basic lsp setup
Using Mason as a package manager -- but at the moment, am planning to
use it for linting/formatting fallback mostly.

I would kind of like to have the Mason install in it's own Lua file, but
following the patterin of @thewatts, a multi-plugin lsp setup is done
within one file. I don't know enough to copy his example with
modifications, and I understand why he's done it this way, so it's good
for now. (and, if I forget... perhaps ever :upside-down-face:)